Stdbts2:
- Split up views into their own files.
- Attempting to make bounties work with bounty_id and contract_version.
