+++ bounties_api/requirements.txt
@@ -1,4 +1,4 @@
- Django>=1.8,<2.0
+ django>=1.8,<2.0
  django-filter==1.1.0
  psycopg2
  redis==2.10.6
