+++ bounties_api/std_bounties/urls.py
@@ -16,7 +16,7 @@ urlpatterns = [
     url(r'^leaderboard/issuer/$', views.LeaderboardIssuer.as_view()),
     url(r'^leaderboard/fulfiller/$', views.LeaderboardFulfiller.as_view()),
     url(r'^token/$', views.Tokens.as_view()),
-    url(r'^bounty/(?P<bounty_id>\d+)/fulfillment/(?P<fulfillment_id>\d+)/review/$', views.SubmissionReviews.as_view()),
+    url(r'^bounty/(?P<bounty_id>\d+)/fulfillment/(?P<fulfillment_id>\d+)/review/$', views.FulfillmentReviews.as_view()),
     url(r'^application/(?P<pk>\d+)/$', views.FulfillerApplicationApplicantView.as_view()),
     url(r'^', include(router.urls))
 ]
