+++ bounties_api/std_bounties/views/comment_views.py
@@ -1,0 +1,39 @@
+from rest_framework.mixins import ListModelMixin
+from rest_framework.viewsets import GenericViewSet
+from django.shortcuts import get_object_or_404
+from django.http import JsonResponse
+from std_bounties.models import Comment, Bounty
+from std_bounties.serializers import CommentSerializer
+from user.permissions import AuthenticationPermission
+from notifications.notification_client import NotificationClient
+notification_client = NotificationClient()
+
+
+class BountyComments(ListModelMixin, GenericViewSet):
+    serializer_class = CommentSerializer
+
+    def get_permissions(self):
+        permission_classes = []
+        if self.request.method == 'POST':
+            permission_classes = [AuthenticationPermission]
+        return [permission() for permission in permission_classes]
+
+    def get_queryset(self):
+        return Comment.objects.filter(bounty__id=self.kwargs['bounty_id']).order_by('-created')
+
+    @staticmethod
+    def post(request, bounty_id):
+        bounty = get_object_or_404(Bounty, bounty_id=bounty_id)
+        serializer = CommentSerializer(
+            data=request.data,
+            context={
+                'request': request
+            })
+        serializer.is_valid(raise_exception=True)
+        comment = serializer.save()
+        bounty.comments.add(comment)
+        notification_client.comment_issued(
+            bounty.bounty_id, comment.created, comment.id)
+        notification_client.comment_received(
+            bounty.bounty_id, comment.created, comment.id)
+        return JsonResponse(serializer.data)
